+++ test/Unit/Ads/CampaignRefreshManagerTests.ts
@@ -427,7 +427,7 @@ describe('CampaignRefreshManager', () => {
                 assert.equal(adsConfig.getPlacement('premium').getState(), PlacementState.READY);
                 assert.equal(adsConfig.getPlacement('video').getState(), PlacementState.READY);
 
-                campaignRefreshManager.setCurrentAdUnit(currentAdUnit);
+                campaignRefreshManager.setCurrentAdUnit(currentAdUnit, placement);
                 currentAdUnit.onStart.trigger();
 
                 assert.equal(campaignRefreshManager.getCampaign('premium'), undefined);
@@ -464,7 +464,7 @@ describe('CampaignRefreshManager', () => {
                 assert.equal(adsConfig.getPlacement('premium').getState(), PlacementState.READY);
                 assert.equal(adsConfig.getPlacement('video').getState(), PlacementState.READY);
 
-                campaignRefreshManager.setCurrentAdUnit(currentAdUnit);
+                campaignRefreshManager.setCurrentAdUnit(currentAdUnit, placement);
                 currentAdUnit.onStart.trigger();
 
                 assert.equal(campaignRefreshManager.getCampaign('premium'), undefined);