+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Tue Jun 11 02:36:41 UTC 2019
+
+* Fix Load Metrics [#10075](https://github.com/Applifier/unity-ads-webview/pull/10075)
+
+# Mon Jun 10 23:42:26 UTC 2019
+
+* Metadata Enabled Per Placement Load Functionality Test [#9685](https://github.com/Applifier/unity-ads-webview/pull/9685)
+
 # Fri Jun  7 19:56:54 UTC 2019
 
 * Consent CTA ab test (bolder fonts) [#10049](https://github.com/Applifier/unity-ads-webview/pull/10049)